Deprecate remaining public scheduler types.
